JB text changes on figure titles:
"Please observe the N of each result to note variables with fewer data points. Information is incomplete for the patients who are still being treated."
-Shouldn't "N" be "total number of patients"? I had changed it in a previous version, but maybe N is better.

"Figure 13: Cumulative patient numbers and outcomes by epidemiological week (of 2020) of admission (or, for patients infected in hospital, of symptom onset). The rightmost bar, marked with an asterisk, represents an incomplete week (due to the 14-day cutoff)."
-The last sentence should be removed as this is no longer in the chart. Correct?

"Figure 14: (a) Treatments used. This only includes patients for whom this information was recorded. (b) The distribution of combinations of antimicrobial treatments and steroids administered during hospital stay, across all patients with completed hospital stay and recorded treatment data. Filled and empty circles below the x-axis indicate treatments that were and were not administered."
-The last sentence should be removed as this is no longer in the chart. Correct?